Fix address sanitizer stack-use-after-scope
std::make_tuple here will copy the arguments into a tuple so it creates
a copy of SmallVector subsliceOffsets and then passes back a tuple with
an ArrayRef. The SmallVector object is then out of scope. Bypassing
make_tuple means that it uses the underlying AllocationSlice's reference
to subsliceOffsets rather than the temporary copy created by make_tuple.