[Backend] Move TMA index translation from mid-end to lowerings (#9082)
Currently the behavior of fp4_padded is different between
`triton::Descriptor` ops and `AsyncTMA` ops. The former is indexed as if
the data is int8, while the latter is indexed by individual fp4
elements, which is what the TMA hardware expects.

This now gets leaked into gluon, which isn't ideal. So, this PR moves
the translation into the lowerings. Along the way, this probably fixes
quite a few bugs as there were several places the translation was
missing.