[AMD] Enable AsyncCopy by default for gfx950 and gfx1250 (#9087)
Enables `ttg.async_copy_global_to_local` for pipelined loads by default
on `gfx950` and `gfx1250`.

This increases LDS consumption because we replace one register buffer
with an additional LDS buffer. After this change, the number of LDS
buffers is equal to `num_stages` (previously it was `num_stages - 1`).
Therefore, some test configs need to be skipped because we run out of
shared memory capacity on `gfx950`.
---------

Co-authored-by: Lei Zhang <antiagainst@gmail.com>