[LinearLayout][tt.scan] Ignore broadcast bases in contig helpers
The logic to find contiguity extent fails in the presence of zero basis
vectors (due to broadcast dims). Add a handleBroadcastDims option to
LinearEncodingAttr contig helpers and use it in scan lowering so broadcasted
register bases don’t distort the axis/non-axis contiguity or elements,
stride, block counts.

Add a unit test that builds a scan with a broadcasted register base and
verify ScanLoweringHelper returns the expected axis/non-axis specific
relevant counts. The observed behavior is difficult to reproduce in LIT tests
without causing another crash mentioned in #7817, hence adding a gtest-based
unit test to keep this solution separate from the next fix for the issue (to
help with the review and feedback addressal process).